# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for CruiseBooking",
        "title": "CruiseBooking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/facilities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Facility"
                ],
                "summary": "List facilities by cruise",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cruise ID",
                        "name": "cruise_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facility"
                ],
                "summary": "Create a facility",
                "parameters": [
                    {
                        "description": "Facility data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FacilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/facilities/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Facility"
                ],
                "summary": "Get a facility",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Facility ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facility"
                ],
                "summary": "Update a facility",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Facility ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Facility data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FacilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Facility"
                ],
                "summary": "Delete a facility",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Facility ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/facility-categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FacilityCategory"
                ],
                "summary": "List all facility categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FacilityCategory"
                ],
                "summary": "Create a facility category",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FacilityCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/facility-categories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FacilityCategory"
                ],
                "summary": "Update a facility category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FacilityCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "FacilityCategory"
                ],
                "summary": "Delete a facility category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cruisebooking_backend_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码（0 表示成功）",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.FacilityCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "icon": {
                    "description": "分类图标",
                    "type": "string"
                },
                "name": {
                    "description": "分类名称（必填）",
                    "type": "string"
                },
                "sort_order": {
                    "description": "排序权重",
                    "type": "integer"
                },
                "status": {
                    "description": "分类状态",
                    "type": "integer"
                }
            }
        },
        "internal_handler.FacilityRequest": {
            "type": "object",
            "required": [
                "category_id",
                "cruise_id",
                "name"
            ],
            "properties": {
                "category_id": {
                    "description": "设施分类 ID（必填）",
                    "type": "integer"
                },
                "charge_price_tip": {
                    "description": "收费提示",
                    "type": "string"
                },
                "cruise_id": {
                    "description": "所属邮轮 ID（必填）",
                    "type": "integer"
                },
                "description": {
                    "description": "设施描述",
                    "type": "string"
                },
                "english_name": {
                    "description": "英文名称",
                    "type": "string"
                },
                "extra_charge": {
                    "description": "是否额外收费",
                    "type": "boolean"
                },
                "location": {
                    "description": "设施位置",
                    "type": "string"
                },
                "name": {
                    "description": "设施名称（必填）",
                    "type": "string"
                },
                "open_hours": {
                    "description": "开放时间",
                    "type": "string"
                },
                "sort_order": {
                    "description": "排序权重",
                    "type": "integer"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "target_audience": {
                    "description": "适合人群",
                    "type": "string"
                }
            }
        },
...
    }
}